add net-cmd helm chart
This helm chart executes a series of shell commands inside
a busybox container. The container has hostNetwork:true and
the NET_ADMIN capability, so it can modify the network
settings of the Kubernetes node.

The commands are run on every node in the cluster because
the container is deployed as a DaemonSet.